[wip] supported v-model in Weex
