+++ src/platforms/weex/compiler/directives/index.js
@@ -1,2 +1,5 @@
+import model from './model'
+
 export default {
+  model
 }
